+++ src/test/java/StatservisTest.java
@@ -7,8 +7,8 @@ public class StatservisTest {
     public void MonthofBigSales() {
         StatsService service = new StatsService();
         long[] sum = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 8;
-        long actual = service.peakSels(sum);
+        int expected = 8;
+        int actual = service.peakSels(sum);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -36,8 +36,8 @@ public class StatservisTest {
     public void MonthofminSales() {
         StatsService service = new StatsService();
         long[] summ = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 9;
-        long actual = service.minSels(summ);
+        int expected = 9;
+        int actual = service.minSels(summ);
 
         Assertions.assertEquals(expected, actual);
     }
